@@ -4,7 +4,7 @@
 requires = ["setuptools", "wheel"]
 
 [tool.black]
-target-version = ['py35', 'py36', 'py37', 'py38']
+target-version = ['py36', 'py37', 'py38']
 
 [tool.isort]
 profile = "black"
