@@ -2,7 +2,7 @@
 #
 # Automatically generated by nengo-bones, do not edit this file directly
 
-import os
+import pathlib
 
 import keras_lmu
 
@@ -96,11 +96,11 @@ release = keras_lmu.__version__  # Full version, with tags
 templates_path = ["_templates"]
 html_static_path = ["_static"]
 html_theme = "nengo_sphinx_theme"
-html_title = "KerasLMU {0} docs".format(release)
+html_title = f"KerasLMU {release} docs"
 htmlhelp_basename = "KerasLMU"
 html_last_updated_fmt = ""  # Default output format (suppressed)
 html_show_sphinx = False
-html_favicon = os.path.join("_static", "favicon.ico")
+html_favicon = str(pathlib.Path("_static", "favicon.ico"))
 html_theme_options = {
     "nengo_logo": "general-full-light.svg",
     "nengo_logo_color": "#a8acaf",
